Run a code in a docker image

Run a code as a string in a docker image

docker_run_code(code, docker_image, volumes = NULL)

Arguments

code
An expression or string of R code

docker_image
A docker image name

volumes
Volume mapping from host to container

Examples

library("altRnative")

# With string
docker_run_code("a = 1 + 1", "ismailsunni/gnur-3.6.1-debian-geospatial")

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript"   "-e"        "a = 1 + 1"

#> <docker_run_output>
#>   $container:
#>     <docker_container>
#>       id: 70c040acfbc0f16ca037e121af4f4f88c3eb1f869cf12ede4f94d37357bcd5a4
#>       name: confident_goldberg
#> 
#>   $logs:
#>     

# With expression
docker_run_code(expression(1 + 1), "ismailsunni/gnur-3.6.1-debian-geospatial")

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript" "-e"      "1 + 1"  
#> O> [1] 2

#> <docker_run_output>
#>   $container:
#>     <docker_container>
#>       id: 3fb6600a5abda9f2e8420f588cb762f8a8aa2b1b22c02960c6858567c23eeb5a
#>       name: magical_noyce
#> 
#>   $logs:
#>     [1] 2

docker_run_code(expression(a = 1 + 1), "ismailsunni/gnur-3.6.1-debian-geospatial")

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript" "-e"      "1 + 1"  
#> O> [1] 2

#> <docker_run_output>
#>   $container:
#>     <docker_container>
#>       id: 84caeb7c7a523f005ae87ed43e7f0a9d79ee2b56138c770cf305084b99832cb5
#>       name: compassionate_sammet
#> 
#>   $logs:
#>     [1] 2

docker_run_code(expression(install.packages("ctv")), "ismailsunni/gnur-3.6.1-debian-geospatial")

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript"                   "-e"                       
#> [3] "install.packages(\"ctv\")"
#> E> Installing package into ‘/usr/local/lib/R/site-library’
#> E> (as ‘lib’ is unspecified)
#> E> trying URL 'https://mran.microsoft.com/snapshot/2019-10-15/src/contrib/ctv_0.8-5.tar.gz'
#> E> Content type 'application/octet-stream' length 363914 bytes (355 KB)
#> E> ==================================================
#> E> downloaded 355 KB
#> E> 
#> E> * installing *source* package ‘ctv’ ...
#> E> ** package ‘ctv’ successfully unpacked and MD5 sums checked
#> E> ** using staged installation
#> E> ** R
#> E> ** inst
#> E> ** byte-compile and prepare package for lazy loading
#> E> ** help
#> E> *** installing help indices
#> E> ** building package indices
#> E> ** installing vignettes
#> E> ** testing if installed package can be loaded from temporary location
#> E> ** testing if installed package can be loaded from final location
#> E> ** testing if installed package keeps a record of temporary installation path
#> E> * DONE (ctv)
#> E> 
#> E> The downloaded source packages are in
#> E> 	‘/tmp/RtmpwIuL8S/downloaded_packages’

#> <docker_run_output>
#>   $container:
#>     <docker_container>
#>       id: 462b1319541b07a879ee3c0406aa24cf9e68c8e23d1cbd4c674a94aa434e41db
#>       name: dazzling_lalande
#> 
#>   $logs:
#>     Installing package into ‘/usr/local/lib/R/site-library’
#>     (as ‘lib’ is unspecified)
#>     trying URL 'https://mran.microsoft.com/snapshot/2019-10-15/src/contrib/ctv_0.8-5.tar.gz'
#>     Content type 'application/octet-stream' length 363914 bytes (355 KB)
#>     ==================================================
#>     [...truncated 13 lines...]
#>     ** testing if installed package can be loaded from final location
#>     ** testing if installed package keeps a record of temporary installation path
#>     * DONE (ctv)
#>     
#>     The downloaded source packages are in
#>     	‘/tmp/RtmpwIuL8S/downloaded_packages’

# With multiple string
docker_run_code(c("a = 1 + 1", "b = a + 2", "print(b)"), "ismailsunni/gnur-3.6.1-debian-geospatial")

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript"                        "-e"                            
#> [3] "a = 1 + 1; b = a + 2; print(b)"
#> O> [1] 4

#> <docker_run_output>
#>   $container:
#>     <docker_container>
#>       id: db3841d8987b901b3cbcab48142343151418807082c45a7716495b3286874e5a
#>       name: stupefied_ardinghelli
#> 
#>   $logs:
#>     [1] 4

code = c("install.packages('ctv')", "library('ctv')", "available.views()")
docker_run_code(code, "ismailsunni/gnur-3.6.1-debian-geospatial")

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript"                                                   
#> [2] "-e"                                                        
#> [3] "install.packages('ctv'); library('ctv'); available.views()"
#> E> Installing package into ‘/usr/local/lib/R/site-library’
#> E> (as ‘lib’ is unspecified)
#> E> trying URL 'https://mran.microsoft.com/snapshot/2019-10-15/src/contrib/ctv_0.8-5.tar.gz'
#> E> Content type 'application/octet-stream' length 363914 bytes (355 KB)
#> E> ==================================================
#> E> downloaded 355 KB
#> E> 
#> E> * installing *source* package ‘ctv’ ...
#> E> ** package ‘ctv’ successfully unpacked and MD5 sums checked
#> E> ** using staged installation
#> E> ** R
#> E> ** inst
#> E> ** byte-compile and prepare package for lazy loading
#> E> ** help
#> E> *** installing help indices
#> E> ** building package indices
#> E> ** installing vignettes
#> E> ** testing if installed package can be loaded from temporary location
#> E> ** testing if installed package can be loaded from final location
#> E> ** testing if installed package keeps a record of temporary installation path
#> E> * DONE (ctv)
#> E> 
#> E> The downloaded source packages are in
#> E> 	‘/tmp/RtmpcUIyPh/downloaded_packages’
#> O> 
#> O> CRAN Task Views
#> O> ---------------
#> O> Name: Bayesian
#> O> Topic: Bayesian Inference
#> O> Maintainer: Jong Hee Park
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ChemPhys
#> O> Topic: Chemometrics and Computational Physics
#> O> Maintainer: Katharine Mullen
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ClinicalTrials
#> O> Topic: Clinical Trial Design, Monitoring, and Analysis
#> O> Maintainer: Ed Zhang and Harry G. Zhang
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Cluster
#> O> Topic: Cluster Analysis & Finite Mixture Models
#> O> Maintainer: Friedrich Leisch and Bettina Gruen
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Databases
#> O> Topic: Databases with R
#> O> Maintainer: Yuan Tang
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: DifferentialEquations
#> O> Topic: Differential Equations
#> O> Maintainer: Karline Soetaert and Thomas Petzoldt
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Distributions
#> O> Topic: Probability Distributions
#> O> Maintainer: Christophe Dutang, Patrice Kiener
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Econometrics
#> O> Topic: Econometrics
#> O> Maintainer: Achim Zeileis
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Environmetrics
#> O> Topic: Analysis of Ecological and Environmental Data
#> O> Maintainer: Gavin Simpson
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ExperimentalDesign
#> O> Topic: Design of Experiments (DoE) & Analysis of Experimental Data
#> O> Maintainer: Ulrike Groemping
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ExtremeValue
#> O> Topic: Extreme Value Analysis
#> O> Maintainer: Christophe Dutang, Kevin Jaunatre
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Finance
#> O> Topic: Empirical Finance
#> O> Maintainer: Dirk Eddelbuettel
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: FunctionalData
#> O> Topic: Functional Data Analysis
#> O> Maintainer: Fabian Scheipl
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Genetics
#> O> Topic: Statistical Genetics
#> O> Maintainer: Giovanni Montana
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Graphics
#> O> Topic: Graphic Displays & Dynamic Graphics & Graphic Devices & Visualization
#> O> Maintainer: Nicholas Lewin-Koh
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: HighPerformanceComputing
#> O> Topic: High-Performance and Parallel Computing with R
#> O> Maintainer: Dirk Eddelbuettel
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Hydrology
#> O> Topic: Hydrological Data and Modeling
#> O> Maintainer: Sam Zipper, Sam Albers, Ilaria Prosdocimi
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: MachineLearning
#> O> Topic: Machine Learning & Statistical Learning
#> O> Maintainer: Torsten Hothorn
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: MedicalImaging
#> O> Topic: Medical Image Analysis
#> O> Maintainer: Brandon Whitcher
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: MetaAnalysis
#> O> Topic: Meta-Analysis
#> O> Maintainer: Michael Dewey
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: MissingData
#> O> Topic: Missing Data
#> O> Maintainer: Julie Josse, Nicholas Tierney and Nathalie Vialaneix (r-miss-tastic team)
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ModelDeployment
#> O> Topic: Model Deployment with R
#> O> Maintainer: Yuan Tang
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Multivariate
#> O> Topic: Multivariate Statistics
#> O> Maintainer: Paul Hewson
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: NaturalLanguageProcessing
#> O> Topic: Natural Language Processing
#> O> Maintainer: Fridolin Wild, Performance Augmentation Lab (PAL, Department of Computing and Communications Technologies, Oxford Brookes University, UK
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: NumericalMathematics
#> O> Topic: Numerical Mathematics
#> O> Maintainer: Hans W. Borchers
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: OfficialStatistics
#> O> Topic: Official Statistics & Survey Methodology
#> O> Maintainer: Matthias Templ
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Optimization
#> O> Topic: Optimization and Mathematical Programming
#> O> Maintainer: Stefan Theussl, 
#> O> 	Florian Schwendinger, Hans W. Borchers
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Pharmacokinetics
#> O> Topic: Analysis of Pharmacokinetic Data
#> O> Maintainer: Bill Denney
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Phylogenetics
#> O> Topic: Phylogenetics, Especially Comparative Methods
#> O> Maintainer: Brian O'Meara
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Psychometrics
#> O> Topic: Psychometric Models and Methods
#> O> Maintainer: Patrick Mair
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ReproducibleResearch
#> O> Topic: Reproducible Research
#> O> Maintainer: Max Kuhn
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Robust
#> O> Topic: Robust Statistical Methods
#> O> Maintainer: Martin Maechler
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: SocialSciences
#> O> Topic: Statistics for the Social Sciences
#> O> Maintainer: John Fox
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Spatial
#> O> Topic: Analysis of Spatial Data
#> O> Maintainer: Roger Bivand
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: SpatioTemporal
#> O> Topic: Handling and Analyzing Spatio-Temporal Data
#> O> Maintainer: Edzer Pebesma
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Survival
#> O> Topic: Survival Analysis
#> O> Maintainer: Arthur Allignol and Aurelien Latouche
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: TeachingStatistics
#> O> Topic: Teaching Statistics
#> O> Maintainer: Paul Northrop
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: TimeSeries
#> O> Topic: Time Series Analysis
#> O> Maintainer: Rob J Hyndman
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: WebTechnologies
#> O> Topic: Web Technologies and Services
#> O> Maintainer: Scott Chamberlain, Thomas Leeper, Patrick Mair, Karthik Ram, Christopher Gandrud
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: gR
#> O> Topic: gRaphical Models in R
#> O> Maintainer: Soren Hojsgaard
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> 

#> <docker_run_output>
#>   $container:
#>     <docker_container>
#>       id: 6895628d3954c91bcd7ccb5c3035522d2f14225ca52b05b13a82de41720eeec0
#>       name: serene_aryabhata
#> 
#>   $logs:
#>     Installing package into ‘/usr/local/lib/R/site-library’
#>     (as ‘lib’ is unspecified)
#>     trying URL 'https://mran.microsoft.com/snapshot/2019-10-15/src/contrib/ctv_0.8-5.tar.gz'
#>     Content type 'application/octet-stream' length 363914 bytes (355 KB)
#>     ==================================================
#>     [...truncated 217 lines...]
#>     ---------------
#>     Name: gR
#>     Topic: gRaphical Models in R
#>     Maintainer: Soren Hojsgaard
#>     Repository: https://mran.microsoft.com/snapshot/2019-10-15
#>     

# With multiple expressions
# This one is not working, see https://github.com/ismailsunni/altRnative/issues/1
# docker_run_code(c(expression(a = 1 + 1, b = a + 2)), "ismailsunni/gnur-3.6.1-debian-geospatial")
# This one is working
code = expression(install.packages("ctv"), library("ctv"), available.views())
docker_run_code(code, "ismailsunni/gnur-3.6.1-debian-geospatial")

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript"                                                       
#> [2] "-e"                                                            
#> [3] "install.packages(\"ctv\"); library(\"ctv\"); available.views()"
#> E> Installing package into ‘/usr/local/lib/R/site-library’
#> E> (as ‘lib’ is unspecified)
#> E> trying URL 'https://mran.microsoft.com/snapshot/2019-10-15/src/contrib/ctv_0.8-5.tar.gz'
#> E> Content type 'application/octet-stream' length 363914 bytes (355 KB)
#> E> ==================================================
#> E> downloaded 355 KB
#> E> 
#> E> * installing *source* package ‘ctv’ ...
#> E> ** package ‘ctv’ successfully unpacked and MD5 sums checked
#> E> ** using staged installation
#> E> ** R
#> E> ** inst
#> E> ** byte-compile and prepare package for lazy loading
#> E> ** help
#> E> *** installing help indices
#> E> ** building package indices
#> E> ** installing vignettes
#> E> ** testing if installed package can be loaded from temporary location
#> E> ** testing if installed package can be loaded from final location
#> E> ** testing if installed package keeps a record of temporary installation path
#> E> * DONE (ctv)
#> E> 
#> E> The downloaded source packages are in
#> E> 	‘/tmp/Rtmp0Eoey9/downloaded_packages’
#> O> 
#> O> CRAN Task Views
#> O> ---------------
#> O> Name: Bayesian
#> O> Topic: Bayesian Inference
#> O> Maintainer: Jong Hee Park
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ChemPhys
#> O> Topic: Chemometrics and Computational Physics
#> O> Maintainer: Katharine Mullen
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ClinicalTrials
#> O> Topic: Clinical Trial Design, Monitoring, and Analysis
#> O> Maintainer: Ed Zhang and Harry G. Zhang
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Cluster
#> O> Topic: Cluster Analysis & Finite Mixture Models
#> O> Maintainer: Friedrich Leisch and Bettina Gruen
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Databases
#> O> Topic: Databases with R
#> O> Maintainer: Yuan Tang
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: DifferentialEquations
#> O> Topic: Differential Equations
#> O> Maintainer: Karline Soetaert and Thomas Petzoldt
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Distributions
#> O> Topic: Probability Distributions
#> O> Maintainer: Christophe Dutang, Patrice Kiener
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Econometrics
#> O> Topic: Econometrics
#> O> Maintainer: Achim Zeileis
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Environmetrics
#> O> Topic: Analysis of Ecological and Environmental Data
#> O> Maintainer: Gavin Simpson
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ExperimentalDesign
#> O> Topic: Design of Experiments (DoE) & Analysis of Experimental Data
#> O> Maintainer: Ulrike Groemping
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ExtremeValue
#> O> Topic: Extreme Value Analysis
#> O> Maintainer: Christophe Dutang, Kevin Jaunatre
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Finance
#> O> Topic: Empirical Finance
#> O> Maintainer: Dirk Eddelbuettel
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: FunctionalData
#> O> Topic: Functional Data Analysis
#> O> Maintainer: Fabian Scheipl
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Genetics
#> O> Topic: Statistical Genetics
#> O> Maintainer: Giovanni Montana
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Graphics
#> O> Topic: Graphic Displays & Dynamic Graphics & Graphic Devices & Visualization
#> O> Maintainer: Nicholas Lewin-Koh
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: HighPerformanceComputing
#> O> Topic: High-Performance and Parallel Computing with R
#> O> Maintainer: Dirk Eddelbuettel
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Hydrology
#> O> Topic: Hydrological Data and Modeling
#> O> Maintainer: Sam Zipper, Sam Albers, Ilaria Prosdocimi
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: MachineLearning
#> O> Topic: Machine Learning & Statistical Learning
#> O> Maintainer: Torsten Hothorn
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: MedicalImaging
#> O> Topic: Medical Image Analysis
#> O> Maintainer: Brandon Whitcher
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: MetaAnalysis
#> O> Topic: Meta-Analysis
#> O> Maintainer: Michael Dewey
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: MissingData
#> O> Topic: Missing Data
#> O> Maintainer: Julie Josse, Nicholas Tierney and Nathalie Vialaneix (r-miss-tastic team)
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ModelDeployment
#> O> Topic: Model Deployment with R
#> O> Maintainer: Yuan Tang
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Multivariate
#> O> Topic: Multivariate Statistics
#> O> Maintainer: Paul Hewson
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: NaturalLanguageProcessing
#> O> Topic: Natural Language Processing
#> O> Maintainer: Fridolin Wild, Performance Augmentation Lab (PAL, Department of Computing and Communications Technologies, Oxford Brookes University, UK
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: NumericalMathematics
#> O> Topic: Numerical Mathematics
#> O> Maintainer: Hans W. Borchers
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: OfficialStatistics
#> O> Topic: Official Statistics & Survey Methodology
#> O> Maintainer: Matthias Templ
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Optimization
#> O> Topic: Optimization and Mathematical Programming
#> O> Maintainer: Stefan Theussl, 
#> O> 	Florian Schwendinger, Hans W. Borchers
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Pharmacokinetics
#> O> Topic: Analysis of Pharmacokinetic Data
#> O> Maintainer: Bill Denney
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Phylogenetics
#> O> Topic: Phylogenetics, Especially Comparative Methods
#> O> Maintainer: Brian O'Meara
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Psychometrics
#> O> Topic: Psychometric Models and Methods
#> O> Maintainer: Patrick Mair
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: ReproducibleResearch
#> O> Topic: Reproducible Research
#> O> Maintainer: Max Kuhn
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Robust
#> O> Topic: Robust Statistical Methods
#> O> Maintainer: Martin Maechler
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: SocialSciences
#> O> Topic: Statistics for the Social Sciences
#> O> Maintainer: John Fox
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Spatial
#> O> Topic: Analysis of Spatial Data
#> O> Maintainer: Roger Bivand
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: SpatioTemporal
#> O> Topic: Handling and Analyzing Spatio-Temporal Data
#> O> Maintainer: Edzer Pebesma
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: Survival
#> O> Topic: Survival Analysis
#> O> Maintainer: Arthur Allignol and Aurelien Latouche
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: TeachingStatistics
#> O> Topic: Teaching Statistics
#> O> Maintainer: Paul Northrop
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: TimeSeries
#> O> Topic: Time Series Analysis
#> O> Maintainer: Rob J Hyndman
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: WebTechnologies
#> O> Topic: Web Technologies and Services
#> O> Maintainer: Scott Chamberlain, Thomas Leeper, Patrick Mair, Karthik Ram, Christopher Gandrud
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> ---------------
#> O> Name: gR
#> O> Topic: gRaphical Models in R
#> O> Maintainer: Soren Hojsgaard
#> O> Repository: https://mran.microsoft.com/snapshot/2019-10-15
#> O> 

#> <docker_run_output>
#>   $container:
#>     <docker_container>
#>       id: a503d0e49e87cafefc362aaf1eeb9cb6ddd0036cb27e784968b6b4814a435e3b
#>       name: musing_leakey
#> 
#>   $logs:
#>     Installing package into ‘/usr/local/lib/R/site-library’
#>     (as ‘lib’ is unspecified)
#>     trying URL 'https://mran.microsoft.com/snapshot/2019-10-15/src/contrib/ctv_0.8-5.tar.gz'
#>     Content type 'application/octet-stream' length 363914 bytes (355 KB)
#>     ==================================================
#>     [...truncated 217 lines...]
#>     ---------------
#>     Name: gR
#>     Topic: gRaphical Models in R
#>     Maintainer: Soren Hojsgaard
#>     Repository: https://mran.microsoft.com/snapshot/2019-10-15
#>     

# This code below is for running sample, need to set proper directory
# code = expression(setwd('/home/docker/sdsr'), bookdown::render_book('index.Rmd', 'bookdown::gitbook'))
# docker_run_code(code, "ismailsunni/gnur-3.6.1-debian-geospatial", volumes = '/home/ismailsunni/dev/r/sdsr:/home/docker/sdsr')