Run benchmark for R code across R implementation and platform

Run benchmark for R code across R implementation and platform.

benchmarks_code(code, platforms = c("debian", "ubuntu"),
  r_implementations = c("gnu-r", "mro"), volumes = NULL, times = 3,
  pull_image = FALSE, ...)

Arguments

code
An expression or string of R code

platforms
List of platform

r_implementations
List of R implementation

volumes
Volume mapping from host to container. Passed to run_code.

times
How many times the code will be run. Passed to microbenchmark

pull_image
If set to TRUE, the needed docker image will be pulled first.

...
Parameters for microbenchmark

Examples

benchmarks_code(code = "1 + 1", times = 3)

#> [1] "1 + 1"
#> [1] "gnu-r" "mro"  
#> [1] "debian" "ubuntu"
#> [1] "Docker image for gnu-r and ubuntu is not supported"
#> [1] "Docker image for mro and debian is not supported"

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript" "-e"      "1 + 1"  
#> O> [1] 2

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript" "-e"      "1 + 1"  
#> O> 
#> O> 
#> O> Microsoft R Services MKL
#> O> End User License Agreement
#> O> 
#> O> These license terms are an agreement between Microsoft Corporation ("Microsoft" or "us") and you ("User" or "you"). Please read them.  By selecting the "I ACCEPT" box, accessing, or otherwise using the R Services MKL software or documentation, you agree to all the terms of this agreement.  If you do not agree with all of these terms, do not select the "I ACCEPT" box, access, or otherwise use the R Services MKL software or documentation.  User does not become a licensed user of, and is not authorized to use, the R Services MKL software or documentation unless and until it has agreed to be bound by all of these terms.
#> O> 
#> O> R Services MKL is a software package provided for use with Microsoft R Server, Microsoft R Open, and any successors or other software applications released by us ("Licensed Products") that can be used with the Intel® Math Kernel Libraries ("MKL")(for more information, see https://software.intel.com/en-us/intel-mkl).  This software package is referred to herein as the "R Services MKL."  Intel Corporation is referred to as "Intel."
#> O> 
#> O> 1. License.  The R Services MKL is licensed for exclusive use with the Licensed Products. Subject to all of the terms and conditions of this Agreement, Microsoft grants User a non-exclusive, non-transferable, non-sublicensable right to use R Services MKL. R Services MKL must be used as a single integrated software application and may not be separated into its constituent parts. 
#> O> 
#> O> 2. Restrictions. User shall not, directly or indirectly: (i) sublicense, resell, rent, lease, distribute, market, commercialize, or otherwise transfer rights or usage to R Services MKL or any component thereof; (ii) modify or create any derivative works of R Services MKL; (iii) remove or alter any copyright, trademark or proprietary notice in R Services MKL; or (iv) reverse engineer, disassemble, attempt to obtain the source code for, decompile or modify any portion of R Services MKL (except to the extent that such restrictions are not permissible under applicable law). 
#> O> 
#> O> 3. Proprietary Rights.  The R Services MKL is licensed, not sold, by its respective owners and licensors. Microsoft, Intel and their licensors shall own all right, title, and interest to R Services MKL and any intellectual or proprietary rights therein (including but not limited to patents, copyrights, trademarks, service marks, trade names and other trade secret rights).  Except as expressly provided herein, no licenses of any kind are granted hereunder, whether by implication, estoppel, or otherwise, and all rights that are not expressly granted are hereby reserved.  
#> O>  	 
#> O> 4. As Is/Disclaimer of Warranty.  TO THE MAXIMUM EXTENT PERMITTED BY APPLICABLE LAW, EXCEPT AS EXPRESSLY STATED IN THIS AGREEMENT, R SERVICES MKL AS PROVIDED BY MICROSOFT AND ITS LICENSORS IS PROVIDED AND LICENSED "AS IS" WITHOUT WARRANTY OF ANY KIND, EITHER EXPRESS, IMPLIED, STATUTORY OR OTHERWISE, INCLUDING, BUT NOT LIMITED TO THE IMPLIED WARRANTIES OF MERCHANTABILITY, NON-INFRINGEMENT AND FITNESS FOR A PARTICULAR PURPOSE.  MICROSOFT AND ITS LICENSORS DO NOT WARRANT OR GUARANTEE THAT THE R SERVICES MKL SOFTWARE WILL MEET USER'S REQUIREMENTS OR THAT THE USE OF R SERVICES MKL WILL BE UNINTERRUPTED OR ERROR FREE.  SOME JURISDICTIONS DO NOT ALLOW THE EXCLUSION OF IMPLIED WARRANTIES SO THE ABOVE EXCLUSIONS MAY NOT APPLY TO YOU.
#> O> 
#> O> 5. Limitation of Liability.  TO THE MAXIMUM EXTENT PERMITTED BY APPLICABLE LAW, IN NO EVENT WILL MICROSOFT, INTEL OR THEIR LICENSORS BE LIABLE FOR ANY DIRECT, INDIRECT, SPECIAL, INCIDENTAL, CONSEQUENTIAL OR EXEMPLARY DAMAGES ARISING OUT OF OR IN ANY WAY RELATING TO THIS AGREEMENT OR THE USE OF OR INABILITY TO USE R SERVICES MKL INCLUDING, WITHOUT LIMITATION, DAMAGES FOR LOSS OF GOODWILL, WORK STOPPAGE, LOST PROFITS, LOSS OF DATA, COMPUTER FAILURE OR ANY AND ALL OTHER COMMERCIAL DAMAGES OR LOSSES EVEN IF ADVISED OF THE POSSIBILITY THEREOF AND REGARDLESS OF THE LEGAL OR EQUITABLE THEORY (CONTRACT, TORT (INCLUDING NEGLIGENCE) OR OTHERWISE) UPON WHICH THE CLAIM IS BASED.  
#> O> 
#> O> 6. General.  User agrees and acknowledges that neither R Services MKL, or any direct product thereof, may be exported outside the United States or re-exported except as authorized and as permitted by the laws and regulations of the United States and/or the laws and regulations of the jurisdiction (if other than the United States) in which User rightfully obtained R Services MKL.  Except as expressly provided herein, User may not assign its rights or delegate its obligations under this Agreement without the prior written consent of Microsoft; Microsoft may assign its rights under this Agreement without condition.  This Agreement will be binding upon and inure to the benefit of the parties, their successors and permitted assigns.  The laws of the State of Washington govern this Agreement.  If federal jurisdiction exists, the parties consent to exclusive jurisdiction and venue in the federal courts in King County, Washington.  If not, the parties consent to exclusive jurisdiction and venue in the Superior Court of King County, Washington.  The parties agree the United Nations Convention on Contracts for the International Sale of Goods will not apply to this Agreement.  The failure of either party to require performance by the other party of any provision hereof shall not affect the full right to require such performance at any time thereafter; nor shall the waiver by either party of a breach of any provision hereof be taken or held to be a waiver of the provision itself.  If any provision of this Agreement is found void and unenforceable, it will be replaced to the extent possible with a provision that comes closest to the meaning of the original provision.  This Agreement, and the documents referenced in this Agreement, constitute the entire agreement between User and Microsoft relating to its subject matter and all terms herein and supersedes all prior or contemporaneous agreements or understandings.  No other terms and conditions shall apply.  This Agreement may be modified or changed only in writing by authorized representatives of User and Microsoft. 
#> O> 
#> O> Intel is a trademark of Intel Corporation in the U.S. and/or other countries.
#> O> 
#> O> 
#> O> 
#> O> 
#> O> Microsoft R Open is an enhanced distribution of R made available by Microsoft under the GNU General Public License v2.
#> O> 
#> O> R is © the R Foundation for Statistical Computing. For more information on R-related products and services, visit http://r-project.org.
#> O> 
#> O> By accepting this agreement you are choosing to download Microsoft R Open and install it on your machine, and agreeing to accept patches and updates to this software according to your Windows Update preferences.
#> O> 
#> O> 
#> O> [1] 2

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript" "-e"      "1 + 1"  
#> O> [1] 2

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript" "-e"      "1 + 1"  
#> O> [1] 2

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript" "-e"      "1 + 1"  
#> O> 
#> O> 
#> O> Microsoft R Services MKL
#> O> End User License Agreement
#> O> 
#> O> These license terms are an agreement between Microsoft Corporation ("Microsoft" or "us") and you ("User" or "you"). Please read them.  By selecting the "I ACCEPT" box, accessing, or otherwise using the R Services MKL software or documentation, you agree to all the terms of this agreement.  If you do not agree with all of these terms, do not select the "I ACCEPT" box, access, or otherwise use the R Services MKL software or documentation.  User does not become a licensed user of, and is not authorized to use, the R Services MKL software or documentation unless and until it has agreed to be bound by all of these terms.
#> O> 
#> O> R Services MKL is a software package provided for use with Microsoft R Server, Microsoft R Open, and any successors or other software applications released by us ("Licensed Products") that can be used with the Intel® Math Kernel Libraries ("MKL")(for more information, see https://software.intel.com/en-us/intel-mkl).  This software package is referred to herein as the "R Services MKL."  Intel Corporation is referred to as "Intel."
#> O> 
#> O> 1. License.  The R Services MKL is licensed for exclusive use with the Licensed Products. Subject to all of the terms and conditions of this Agreement, Microsoft grants User a non-exclusive, non-transferable, non-sublicensable right to use R Services MKL. R Services MKL must be used as a single integrated software application and may not be separated into its constituent parts. 
#> O> 
#> O> 2. Restrictions. User shall not, directly or indirectly: (i) sublicense, resell, rent, lease, distribute, market, commercialize, or otherwise transfer rights or usage to R Services MKL or any component thereof; (ii) modify or create any derivative works of R Services MKL; (iii) remove or alter any copyright, trademark or proprietary notice in R Services MKL; or (iv) reverse engineer, disassemble, attempt to obtain the source code for, decompile or modify any portion of R Services MKL (except to the extent that such restrictions are not permissible under applicable law). 
#> O> 
#> O> 3. Proprietary Rights.  The R Services MKL is licensed, not sold, by its respective owners and licensors. Microsoft, Intel and their licensors shall own all right, title, and interest to R Services MKL and any intellectual or proprietary rights therein (including but not limited to patents, copyrights, trademarks, service marks, trade names and other trade secret rights).  Except as expressly provided herein, no licenses of any kind are granted hereunder, whether by implication, estoppel, or otherwise, and all rights that are not expressly granted are hereby reserved.  
#> O>  	 
#> O> 4. As Is/Disclaimer of Warranty.  TO THE MAXIMUM EXTENT PERMITTED BY APPLICABLE LAW, EXCEPT AS EXPRESSLY STATED IN THIS AGREEMENT, R SERVICES MKL AS PROVIDED BY MICROSOFT AND ITS LICENSORS IS PROVIDED AND LICENSED "AS IS" WITHOUT WARRANTY OF ANY KIND, EITHER EXPRESS, IMPLIED, STATUTORY OR OTHERWISE, INCLUDING, BUT NOT LIMITED TO THE IMPLIED WARRANTIES OF MERCHANTABILITY, NON-INFRINGEMENT AND FITNESS FOR A PARTICULAR PURPOSE.  MICROSOFT AND ITS LICENSORS DO NOT WARRANT OR GUARANTEE THAT THE R SERVICES MKL SOFTWARE WILL MEET USER'S REQUIREMENTS OR THAT THE USE OF R SERVICES MKL WILL BE UNINTERRUPTED OR ERROR FREE.  SOME JURISDICTIONS DO NOT ALLOW THE EXCLUSION OF IMPLIED WARRANTIES SO THE ABOVE EXCLUSIONS MAY NOT APPLY TO YOU.
#> O> 
#> O> 5. Limitation of Liability.  TO THE MAXIMUM EXTENT PERMITTED BY APPLICABLE LAW, IN NO EVENT WILL MICROSOFT, INTEL OR THEIR LICENSORS BE LIABLE FOR ANY DIRECT, INDIRECT, SPECIAL, INCIDENTAL, CONSEQUENTIAL OR EXEMPLARY DAMAGES ARISING OUT OF OR IN ANY WAY RELATING TO THIS AGREEMENT OR THE USE OF OR INABILITY TO USE R SERVICES MKL INCLUDING, WITHOUT LIMITATION, DAMAGES FOR LOSS OF GOODWILL, WORK STOPPAGE, LOST PROFITS, LOSS OF DATA, COMPUTER FAILURE OR ANY AND ALL OTHER COMMERCIAL DAMAGES OR LOSSES EVEN IF ADVISED OF THE POSSIBILITY THEREOF AND REGARDLESS OF THE LEGAL OR EQUITABLE THEORY (CONTRACT, TORT (INCLUDING NEGLIGENCE) OR OTHERWISE) UPON WHICH THE CLAIM IS BASED.  
#> O> 
#> O> 6. General.  User agrees and acknowledges that neither R Services MKL, or any direct product thereof, may be exported outside the United States or re-exported except as authorized and as permitted by the laws and regulations of the United States and/or the laws and regulations of the jurisdiction (if other than the United States) in which User rightfully obtained R Services MKL.  Except as expressly provided herein, User may not assign its rights or delegate its obligations under this Agreement without the prior written consent of Microsoft; Microsoft may assign its rights under this Agreement without condition.  This Agreement will be binding upon and inure to the benefit of the parties, their successors and permitted assigns.  The laws of the State of Washington govern this Agreement.  If federal jurisdiction exists, the parties consent to exclusive jurisdiction and venue in the federal courts in King County, Washington.  If not, the parties consent to exclusive jurisdiction and venue in the Superior Court of King County, Washington.  The parties agree the United Nations Convention on Contracts for the International Sale of Goods will not apply to this Agreement.  The failure of either party to require performance by the other party of any provision hereof shall not affect the full right to require such performance at any time thereafter; nor shall the waiver by either party of a breach of any provision hereof be taken or held to be a waiver of the provision itself.  If any provision of this Agreement is found void and unenforceable, it will be replaced to the extent possible with a provision that comes closest to the meaning of the original provision.  This Agreement, and the documents referenced in this Agreement, constitute the entire agreement between User and Microsoft relating to its subject matter and all terms herein and supersedes all prior or contemporaneous agreements or understandings.  No other terms and conditions shall apply.  This Agreement may be modified or changed only in writing by authorized representatives of User and Microsoft. 
#> O> 
#> O> Intel is a trademark of Intel Corporation in the U.S. and/or other countries.
#> O> 
#> O> 
#> O> 
#> O> 
#> O> Microsoft R Open is an enhanced distribution of R made available by Microsoft under the GNU General Public License v2.
#> O> 
#> O> R is © the R Foundation for Statistical Computing. For more information on R-related products and services, visit http://r-project.org.
#> O> 
#> O> By accepting this agreement you are choosing to download Microsoft R Open and install it on your machine, and agreeing to accept patches and updates to this software according to your Windows Update preferences.
#> O> 
#> O> 
#> O> [1] 2

#> Detected API version '1.40' is above max version '1.39'; downgrading

#> [1] "Full command"
#> [1] "Rscript" "-e"      "1 + 1"  
#> O> 
#> O> 
#> O> Microsoft R Services MKL
#> O> End User License Agreement
#> O> 
#> O> These license terms are an agreement between Microsoft Corporation ("Microsoft" or "us") and you ("User" or "you"). Please read them.  By selecting the "I ACCEPT" box, accessing, or otherwise using the R Services MKL software or documentation, you agree to all the terms of this agreement.  If you do not agree with all of these terms, do not select the "I ACCEPT" box, access, or otherwise use the R Services MKL software or documentation.  User does not become a licensed user of, and is not authorized to use, the R Services MKL software or documentation unless and until it has agreed to be bound by all of these terms.
#> O> 
#> O> R Services MKL is a software package provided for use with Microsoft R Server, Microsoft R Open, and any successors or other software applications released by us ("Licensed Products") that can be used with the Intel® Math Kernel Libraries ("MKL")(for more information, see https://software.intel.com/en-us/intel-mkl).  This software package is referred to herein as the "R Services MKL."  Intel Corporation is referred to as "Intel."
#> O> 
#> O> 1. License.  The R Services MKL is licensed for exclusive use with the Licensed Products. Subject to all of the terms and conditions of this Agreement, Microsoft grants User a non-exclusive, non-transferable, non-sublicensable right to use R Services MKL. R Services MKL must be used as a single integrated software application and may not be separated into its constituent parts. 
#> O> 
#> O> 2. Restrictions. User shall not, directly or indirectly: (i) sublicense, resell, rent, lease, distribute, market, commercialize, or otherwise transfer rights or usage to R Services MKL or any component thereof; (ii) modify or create any derivative works of R Services MKL; (iii) remove or alter any copyright, trademark or proprietary notice in R Services MKL; or (iv) reverse engineer, disassemble, attempt to obtain the source code for, decompile or modify any portion of R Services MKL (except to the extent that such restrictions are not permissible under applicable law). 
#> O> 
#> O> 3. Proprietary Rights.  The R Services MKL is licensed, not sold, by its respective owners and licensors. Microsoft, Intel and their licensors shall own all right, title, and interest to R Services MKL and any intellectual or proprietary rights therein (including but not limited to patents, copyrights, trademarks, service marks, trade names and other trade secret rights).  Except as expressly provided herein, no licenses of any kind are granted hereunder, whether by implication, estoppel, or otherwise, and all rights that are not expressly granted are hereby reserved.  
#> O>  	 
#> O> 4. As Is/Disclaimer of Warranty.  TO THE MAXIMUM EXTENT PERMITTED BY APPLICABLE LAW, EXCEPT AS EXPRESSLY STATED IN THIS AGREEMENT, R SERVICES MKL AS PROVIDED BY MICROSOFT AND ITS LICENSORS IS PROVIDED AND LICENSED "AS IS" WITHOUT WARRANTY OF ANY KIND, EITHER EXPRESS, IMPLIED, STATUTORY OR OTHERWISE, INCLUDING, BUT NOT LIMITED TO THE IMPLIED WARRANTIES OF MERCHANTABILITY, NON-INFRINGEMENT AND FITNESS FOR A PARTICULAR PURPOSE.  MICROSOFT AND ITS LICENSORS DO NOT WARRANT OR GUARANTEE THAT THE R SERVICES MKL SOFTWARE WILL MEET USER'S REQUIREMENTS OR THAT THE USE OF R SERVICES MKL WILL BE UNINTERRUPTED OR ERROR FREE.  SOME JURISDICTIONS DO NOT ALLOW THE EXCLUSION OF IMPLIED WARRANTIES SO THE ABOVE EXCLUSIONS MAY NOT APPLY TO YOU.
#> O> 
#> O> 5. Limitation of Liability.  TO THE MAXIMUM EXTENT PERMITTED BY APPLICABLE LAW, IN NO EVENT WILL MICROSOFT, INTEL OR THEIR LICENSORS BE LIABLE FOR ANY DIRECT, INDIRECT, SPECIAL, INCIDENTAL, CONSEQUENTIAL OR EXEMPLARY DAMAGES ARISING OUT OF OR IN ANY WAY RELATING TO THIS AGREEMENT OR THE USE OF OR INABILITY TO USE R SERVICES MKL INCLUDING, WITHOUT LIMITATION, DAMAGES FOR LOSS OF GOODWILL, WORK STOPPAGE, LOST PROFITS, LOSS OF DATA, COMPUTER FAILURE OR ANY AND ALL OTHER COMMERCIAL DAMAGES OR LOSSES EVEN IF ADVISED OF THE POSSIBILITY THEREOF AND REGARDLESS OF THE LEGAL OR EQUITABLE THEORY (CONTRACT, TORT (INCLUDING NEGLIGENCE) OR OTHERWISE) UPON WHICH THE CLAIM IS BASED.  
#> O> 
#> O> 6. General.  User agrees and acknowledges that neither R Services MKL, or any direct product thereof, may be exported outside the United States or re-exported except as authorized and as permitted by the laws and regulations of the United States and/or the laws and regulations of the jurisdiction (if other than the United States) in which User rightfully obtained R Services MKL.  Except as expressly provided herein, User may not assign its rights or delegate its obligations under this Agreement without the prior written consent of Microsoft; Microsoft may assign its rights under this Agreement without condition.  This Agreement will be binding upon and inure to the benefit of the parties, their successors and permitted assigns.  The laws of the State of Washington govern this Agreement.  If federal jurisdiction exists, the parties consent to exclusive jurisdiction and venue in the federal courts in King County, Washington.  If not, the parties consent to exclusive jurisdiction and venue in the Superior Court of King County, Washington.  The parties agree the United Nations Convention on Contracts for the International Sale of Goods will not apply to this Agreement.  The failure of either party to require performance by the other party of any provision hereof shall not affect the full right to require such performance at any time thereafter; nor shall the waiver by either party of a breach of any provision hereof be taken or held to be a waiver of the provision itself.  If any provision of this Agreement is found void and unenforceable, it will be replaced to the extent possible with a provision that comes closest to the meaning of the original provision.  This Agreement, and the documents referenced in this Agreement, constitute the entire agreement between User and Microsoft relating to its subject matter and all terms herein and supersedes all prior or contemporaneous agreements or understandings.  No other terms and conditions shall apply.  This Agreement may be modified or changed only in writing by authorized representatives of User and Microsoft. 
#> O> 
#> O> Intel is a trademark of Intel Corporation in the U.S. and/or other countries.
#> O> 
#> O> 
#> O> 
#> O> 
#> O> Microsoft R Open is an enhanced distribution of R made available by Microsoft under the GNU General Public License v2.
#> O> 
#> O> R is © the R Foundation for Statistical Computing. For more information on R-related products and services, visit http://r-project.org.
#> O> 
#> O> By accepting this agreement you are choosing to download Microsoft R Open and install it on your machine, and agreeing to accept patches and updates to this software according to your Windows Update preferences.
#> O> 
#> O> 
#> O> [1] 2

#> Unit: seconds
#>             expr      min       lq     mean   median       uq      max neval
#>  gnu-r on debian 1.480605 1.669847 1.806470 1.859089 1.969403 2.079717     3
#>    mro on ubuntu 1.136482 1.226839 1.643258 1.317197 1.896646 2.476095     3

# This code below is for running sample, need to set proper directory
# code = expression(setwd('/home/docker/sdsr'), bookdown::clean_book(TRUE), unlink('_book/', recursive=TRUE), unlink('_bookdown_files', recursive=TRUE), bookdown::render_book('index.Rmd', 'bookdown::gitbook'))
# benchmarks_code(code = code, volumes = '/home/ismailsunni/dev/r/sdsr:/home/docker/sdsr', times = 3)